src/main/java/challenges/solution/BinaryTreeHeight.java
package challenges.solution;

class Node {
    Node left;
    Node right;
    int data;

    Node(int data) {
        this.data = data;
        left = null;
        right = null;
    }
}

public class BinaryTreeHeight {
    public static int height(Node root) {
        if(root == null)
            return -1;
        int left = 1 + height(root.left);
        int right = 1 + height(root.right);

        return Math.max(left, right);
    }

    public static Node insert(Node root, int data) {
        if (root == null) {
            return new Node(data);
        } else {
            Node cur;
            if (data <= root.data) {
                cur = insert(root.left, data);
                root.left = cur;
            } else {
                cur = insert(root.right, data);
                root.right = cur;
            }
            return root;
        }
    }
}
